Fix handling of android events
The way we are passing events along in the android side, the callbacks
would receive and event object.  However, the interface expects the
underlying nativeEvent (i.e. a map with the fields as on iOS), so we
need to put together simple wrappers for our events to make things
match.

Fixes #29