Support on Android for Twilio Video v2
This updates the interface to work with Twilio Video v2.1.  In addition to a bunch of general cleanup, it updates the events to be more similar to their iOS counterparts.  The same user changes needed to get things working on iOS should also work here (namely, the use of Sids to identify everything).